refactor: extract DelegateGigPage components to separate files
